+++ src/main/java/com/linkedin/learning/repository/RoomRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.repository.CrudRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface RoomRepository extends CrudRepository<RoomEntity, Long> {
+public interface RoomRepository extends CrudRepository<RoomEntity, Integer> {
     @Override
-    Optional<RoomEntity> findById(Long id);
+    Optional<RoomEntity> findById(Integer id);
     @Override
     List<RoomEntity> findAll();
 }
